@@ -1,5 +1,10 @@
-﻿using CC.Web.Api.Filter;
+﻿using Autofac;
+using Autofac.Extensions.DependencyInjection;
+using CC.Web.Api.Core;
+using CC.Web.Api.Filter;
 using CC.Web.Dao;
+using CC.Web.Dao.Core;
+using CC.Web.Service.System;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
@@ -7,6 +12,8 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using ServiceStack.Redis;
+using System;
+using System.Reflection;
 
 namespace CC.Web.Api
 {
@@ -20,18 +27,28 @@ namespace CC.Web.Api
         public IConfiguration Configuration { get; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
+        public IServiceProvider ConfigureServices(IServiceCollection services)
         {
             services.AddSingleton<IRedisClientsManager>(new BasicRedisClientManager());
             services.AddMvc(option => option.Filters.Add(typeof(ActionLogFilter)))
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
-            services.AddDbContext<CCDbContext>(option => option.UseSqlServer(Configuration.GetConnectionString("CCDatabase")));
+                .SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
+            services.AddDbContext<CCDbContext>(option => 
+            option.UseSqlServer(Configuration.GetConnectionString("CCDatabase"), b => b.MigrationsAssembly("CC.Web.Api")));
+
+            services.AddScoped(typeof(IRepository<>), typeof(Repository<>));
+            services.AddSingleton<IUserService, UserService>();
+
+            var containerBuilder = new ContainerBuilder();
+            containerBuilder.Populate(services);
+            AutofacConfig.RegisterObj(containerBuilder);
+            var applicationContainer = containerBuilder.Build();
+            return new AutofacServiceProvider(applicationContainer);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            if (env.IsDevelopment())
+            if (env.EnvironmentName == "Development")
             {
                 app.UseDeveloperExceptionPage();
             }
@@ -42,7 +59,15 @@ namespace CC.Web.Api
             }
 
             app.UseHttpsRedirection();
-            app.UseMvc();
+
+            app.UseRouting();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllerRoute(
+                    name: "default",
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
+                endpoints.MapRazorPages();
+            });
         }
     }
 }
