@@ -1,13 +1,14 @@
 ﻿using CC.Web.Model.System;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace CC.Web.Dao
+namespace CC.Web.Dao.Core
 {
     public class CCDbContext : DbContext
     {
+        public CCDbContext(DbContextOptions<CCDbContext> options) : base(options) 
+        { 
+        }
+
         public DbSet<User> Users { get; set; }
     }
 }
